Swap around `%x`, `%X`, `%h`, and `%H`.
Within SA:MP basically everyone uses `%x` for formatting HEX characters, and that used upper-case letters for `A`-`F`.  The old version of this code changed that behaviour to use lower-case letters and added `%X` for upper-case letters.  Being able to specify lower- or upper-case is good, but we should stay backwards-compatible with as much other code as possible.  This also had `%h` but no `%H` so I switched `%x` to keep historical behaviour and instead use `%h` and `%H` for case specification.