Change priority again
I previously changed the order to:

* Components
* Plugins
* Defaults

The theory being that then components could override any other natives from other places, which is still a good thing.  However, in SA:MP plugins have lower priority than the default natives and it turns out that at least one plugin (sscanf) implicitly relies on this order.  Given that they are legacy anyway, the order should be:

* Components
* Defaults
* Plugins

This does that.

It also calls `OnGameModeExit` in filterscripts when the pawn component shuts down.