Fix custom rules getting overridden by automatic values every tick.
This keeps a set of keys that have had a custom value given, and if a key is found in that set only a new custom value can be assigned to it - it can't be auto-assigned.  Deleting a variable is also customising it so this flag is set in that case as well, which gives the interesting property that we can find deleted keys by checking if they are customised but don't exist.

Also, adding queries used to use a templated function with variable arguments, but then immediately assumed the parameters were convertible to strings.  There was nothing to enforce this in the signature and none of the calls took advantage of this in any way so I just removed it and used two `StringView` parameters.