Extend the query rules API a lot:
`ResetServerRule` - Used after setting a custom value to return the rule to an automatically set state.
`IsCustomServerRule` - `true` for rules set by script, `false` for those set by the server.
`IsRemovedServerRule` - `true` for deleted rules.
`IsProtectedServerRule` - `true` for rules that can't be modified.
`ProtectServerRule` - Mark a rule as protected.  Once set can never be unset (that would defeat the object of protection).
`GetServerRule` - Return the string value of a rule.