Separate `cab` and `trailer`:
Having them in a union together really doesn't save that much space, since there's a `towing` variable required to differentiate the cases, plus the resulting code becomes a lot more complicated.  There's also an argument to be had that being a trailer and a cab are not mutually exclusive - trains have both, baggage carts have both.